examples/scaffold/pyutil/requirements.txt
# **NOTE**
# If you edit this file then delete the `requirements.lock` 
# file and run `./cli.py` to update the new dependencies

PyYAML
yaclipy-tools
